Partial Commit

+++ SimpleService/Models/Register.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Web;
 
@@ -10,7 +12,7 @@ namespace SimpleService.Models
     /// </summary>
     public class Register
     {
-        public byte Id { get; set; }
+        public int Id { get; set; }
         
         public string Left { get; set; }
